backend/ttw_backend/requirements.txt
alembic==1.15.2
altair==5.5.0
amqp==5.3.1
asgiref==3.11.0
attrs==25.3.0
autobahn==25.11.1
Automat==25.4.16
beautifulsoup4==4.13.4
billiard==4.2.4
blinker==1.9.0
cachetools==5.5.2
cbor2==5.7.1
celery==5.6.0
certifi==2025.4.26
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.6.1
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.2
narwhals==1.42.1
numpy==2.3.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
playwright==1.52.0
prompt_toolkit==3.0.52
protobuf==6.31.1
psycopg2-binary==2.9.10
py-ubjson==0.16.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
pyee==13.0.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
redis==7.1.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
streamlit==1.45.1
stripe==14.0.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.11.0
zope.interface==8.1.1
